Add `backup::create`
This new function creates a new RocksDB backup. In a future release,
this function will be enhanced to support creation of PostgreSQL backups
as well. This provides a centralized and consistent interface for
creating backups across different types of databases.
